Updated Piece script, Finished chess rules and legal moves, integrated checkmate
Piece script is now not a gameobject, but only stores info about the piece. The previous object is now pieceSprite, which controls the piece image.
Legal moves are processed before moves, however if the move leaves you in check, the move will be undone and not added to notation. If a move puts the opponent in checkmate, variables are updated and a '#' is added to notation