Integrated castling, and FEN, working on EnPassant
Each move creates a FEN object that is attached to the notation. Will add a feature to click on the notation to show the board state after that move. Castling also integrated, including not being able to castle through check. EnPassant works as far as the code knows when it is legal, need to remove the taken piece from enpassant and make sure it is notated correctly